+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## 1.1.1 under development
 
-- no changes in this release.
+- New #65: Add `IP_PATTERN` and `IP_REGEXP` constants  to `IpHelper` for checking IP of both IPv4 and IPv6 versions             
+  (@arogachev)
+- New #65: Add `NEGATION_CHARACTER` constant to `IpRanges` used to negate ranges (@arogachev)
+- New #65: Add `isIpv4()`, `isIpv6()`, `isIp()` methods to `IpHelper` (@arogachev)
 
 ## 1.1.0 August 06, 2024